requirements.txt
Django==1.4.10
South==0.8.4
argparse==1.2.1
django-dajax==0.9.2
django-dajaxice==0.5.5
django-grappelli==2.4.8
django-registration==1.0
django-sslserver==0.11
jsonpickle==0.6.1
simplejson==3.3.2
wsgiref==0.1.2
